chore(deps-dev): deps updates, remove old one, refactor docs
