+++ package.json
@@ -2,112 +2,118 @@
   "name": "ember-steps",
   "version": "10.0.1",
   "description": "Ember addon for declaratively creating wizards, tabbed UIs, and more",
+  "keywords": [
+    "ember-addon"
+  ],
+  "homepage": "https://alexlafroscia.github.io/ember-steps",
+  "repository": "https://github.com/alexlafroscia/ember-steps",
+  "license": "MIT",
+  "author": "Alex LaFroscia <alex@lafroscia.com>",
+  "main": "index.js",
+  "directories": {
+    "doc": "doc",
+    "test": "tests"
+  },
   "scripts": {
     "build": "ember build",
+    "postinstall": "husky install",
     "lint:hbs": "ember-template-lint .",
     "lint:js": "eslint .",
+    "prepublishOnly": "pinst --disable",
+    "postpublish": "pinst --enable",
+    "release": "standard-version",
     "start": "ember serve",
     "test": "ember test",
-    "test:all": "ember try:each",
-    "release": "standard-version",
-    "postinstall": "husky install",
-    "prepublishOnly": "pinst --disable",
-    "postpublish": "pinst --enable"
+    "test:all": "ember try:each"
   },
-  "main": "index.js",
-  "homepage": "https://alexlafroscia.github.io/ember-steps",
-  "engines": {
-    "node": "10.* || >= 12"
+  "lint-staged": {
+    "*.js": "eslint --fix",
+    "*.{md,json,yml}": "prettier --write"
   },
   "dependencies": {
-    "@glimmer/component": "^1.0.0",
-    "@glimmer/tracking": "^1.0.0",
+    "@glimmer/component": "^1.1.2",
+    "@glimmer/tracking": "^1.1.2",
     "ember-cli-babel": "^7.26.8",
     "ember-cli-htmlbars": "^6.0.1",
     "tracked-maps-and-sets": "^3.0.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "^13.1.0",
-    "@commitlint/config-conventional": "^13.1.0",
+    "@commitlint/cli": "^17.1.2",
+    "@commitlint/config-conventional": "^17.1.0",
     "@ember/optional-features": "^2.0.0",
     "@ember/test-helpers": "^2.6.0",
-    "@embroider/test-setup": "^0.48.1",
+    "@embroider/test-setup": "^1.8.3",
     "babel-eslint": "^10.1.0",
     "broccoli-asset-rev": "^3.0.0",
     "ember-auto-import": "^2.2.4",
     "ember-cli": "~4.0.1",
-    "ember-cli-addon-docs": "^3.0.0",
-    "ember-cli-addon-docs-esdoc": "^0.4.0",
+    "ember-cli-addon-docs": "^5.0.0",
+    "ember-cli-addon-docs-yuidoc": "^1.0.0",
     "ember-cli-dependency-checker": "^3.2.0",
     "ember-cli-deploy": "^1.0.2",
     "ember-cli-deploy-build": "^2.0.0",
     "ember-cli-deploy-git": "^1.3.4",
     "ember-cli-deploy-git-ci": "^1.0.1",
-    "ember-cli-document-title-northm": "^1.0.3",
-    "ember-cli-fastboot": "^2.2.3",
-    "ember-cli-fastboot-testing": "^0.5.0",
+    "ember-cli-fastboot": "^3.3.2",
+    "ember-cli-fastboot-testing": "^0.6.0",
     "ember-cli-inject-live-reload": "^2.0.2",
     "ember-cli-sri": "^2.1.1",
     "ember-cli-uglify": "^3.0.0",
-    "ember-composable-helpers": "^4.2.0",
-    "ember-concurrency": "^2.0.0",
+    "ember-composable-helpers": "^5.0.0",
+    "ember-concurrency": "^2.3.7",
     "ember-concurrency-decorators": "^2.0.0",
+    "ember-data": "^3.28.7",
     "ember-disable-prototype-extensions": "^1.1.3",
     "ember-export-application-global": "^2.0.1",
-    "ember-fetch": "^8.0.1",
+    "ember-fetch": "^8.1.2",
     "ember-load-initializers": "^2.1.1",
-    "ember-qunit": "^5.1.5",
+    "ember-qunit": "^6.0.0",
     "ember-qunit-assert-helpers": "^0.2.2",
     "ember-resolver": "^8.0.3",
     "ember-source": "~4.0.1",
     "ember-source-channel-url": "^3.0.0",
-    "ember-template-lint": "^3.15.0",
-    "ember-truth-helpers": "^3.0.0",
+    "ember-template-lint": "^4.15.0",
+    "ember-truth-helpers": "^3.1.1",
     "ember-try": "^2.0.0",
     "eslint": "^7.2.0",
-    "eslint-config-prettier": "^8.0.0",
-    "eslint-plugin-ember": "^10.5.8",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-ember": "^11.0.6",
     "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-qunit": "^7.3.1",
-    "husky": "^7.0.1",
-    "lint-staged": "^11.0.0",
+    "husky": "^8.0.1",
+    "lint-staged": "^13.0.3",
     "loader.js": "^4.7.0",
-    "pinst": "^2.1.4",
+    "pinst": "^3.0.0",
     "prettier": "^2.5.1",
     "qunit": "^2.17.2",
     "qunit-dom": "^2.0.0",
-    "qunit-wait-for": "^2.0.0",
-    "standard-version": "^9.0.0",
+    "qunit-wait-for": "^2.0.2",
+    "standard-version": "^9.5.0",
     "testdouble": "^3.16.0",
     "testdouble-qunit": "^3.0.0",
     "webpack": "^5.65.0"
   },
-  "keywords": [
-    "ember-addon"
-  ],
-  "repository": "https://github.com/alexlafroscia/ember-steps",
-  "license": "MIT",
-  "author": "Alex LaFroscia <alex@lafroscia.com>",
-  "directories": {
-    "doc": "doc",
-    "test": "tests"
+  "engines": {
+    "node": "10.* || >= 12"
+  },
+  "volta": {
+    "node": "14.20.1",
+    "yarn": "1.22.19"
+  },
+  "ember": {
+    "edition": "octane"
   },
   "ember-addon": {
     "configPath": "tests/dummy/config",
     "demoURL": "https://alexlafroscia.github.io/ember-steps"
-  },
-  "lint-staged": {
-    "*.js": "eslint --fix",
-    "*.{md,json,yml}": "prettier --write"
   },
   "standard-version": {
     "scripts": {
       "postchangelog": "prettier --write CHANGELOG.md"
     }
   },
-  "volta": {
-    "node": "14.20.1",
-    "yarn": "1.22.19"
+  "resolutions": {
+    "ember-cli-babel": "^7.26.8"
   }
 }
