+++ tests/dummy/app/app.js
@@ -5,7 +5,6 @@ import config from './config/environment';
 
 export default class App extends Application {
   modulePrefix = config.modulePrefix;
-  podModulePrefix = config.podModulePrefix;
   Resolver = Resolver;
 }
