--- tests/dummy/app/application/template.hbs
@@ -1,3 +1,0 @@
-{{outlet}}
-
-<DocsKeyboardShortcuts />
