+++ tests/dummy/app/components/docs/cookbook/dynamic-definition/basic-demo.hbs
