+++ tests/dummy/app/components/docs/cookbook/tabs/demo-basic-tabs.hbs
