+++ tests/dummy/app/components/docs/cookbook/wizard/demo-basic-wizard.hbs
