+++ tests/dummy/app/components/docs/cookbook/wizard/demo-wizard-with-progress-indicator.hbs
