+++ tests/dummy/app/components/docs/features/validating-steps/async-password-check.js
