+++ tests/dummy/app/components/docs/features/validating-steps/basic-usage.js
@@ -1,0 +1,16 @@
+// BEGIN-SNIPPET validating-steps-basic-usage.js
+import Component from '@glimmer/component';
+import { action } from '@ember/object';
+import { tracked } from '@glimmer/tracking';
+
+export default class extends Component {
+  @tracked password = '';
+
+  @action
+  checkPassword(resolve) {
+    if (this.password === 'password') {
+      resolve();
+    }
+  }
+}
+// END-SNIPPET
