+++ tests/dummy/app/config/environment.d.ts
@@ -10,7 +10,6 @@ export default config;
 declare const config: {
   environment: any;
   modulePrefix: string;
-  podModulePrefix: string;
   locationType: string;
   rootURL: string;
 };
