--- tests/dummy/app/docs/cookbook/tabs/index/template.md
@@ -1,5 +1,0 @@
-# Tabbed Interface
-
-A tabbed interface is a really common UI pattern that is often implemented using a number of conditionals in the template. You can simplify this greatly using a `step-manager` that tracks which tab is active and allows you to "transition" to a new tab.
-
-{{docs/cookbook/tabs/demo-basic-tabs}}
