--- tests/dummy/app/docs/cookbook/wizard/index/template.md
@@ -1,7 +1,0 @@
-# Basic Wizard
-
-`ember-steps` can be used to create a "wizard" interface that walks users through a series of steps one at a time. In the basic example below, we simply allow the user to advance "forward" and "backward" through the set of steps.
-
-Note that we can prevent the user from "wrapping around" to the first step by disabling the progress button once they get to the last step.
-
-{{docs/cookbook/wizard/demo-basic-wizard}}
