--- tests/dummy/app/docs/cookbook/wizard/progress-indicator/template.md
@@ -1,5 +1,0 @@
-# Wizard with Progress Indicator
-
-What if you want to provide the user with a progress indicator that allows them to represent where they are in the wizard? Since the `step-manager` yields all of the step names as an `Array`, you can dynamically create a progress indicator. You can even allow the user to navigate to a specific step by clicking on the location in the indicator.
-
-{{docs/cookbook/wizard/demo-wizard-with-progress-indicator}}
