--- tests/dummy/app/docs/features/inactive/template.md
@@ -1,5 +1,0 @@
-# Rendering Inactive Steps
-
-In most cases you'll only want to show one step at a time. However, if you would rather display a "disabled" state for steps that are not active, you can do that too!
-
-{{docs/features/inactive/example-1}}
