--- tests/dummy/app/docs/features/state-manager/step-yielded/template.hbs
@@ -1,36 +1,0 @@
-<DocsDemo as |demo|>
-  <demo.example @name="demo-cyclical-steps-step-yielded.hbs">
-    <StepManager as |w|>
-      <w.Step as |step|>
-        <h3>First</h3>
-        There is a first step: {{step.hasNext}} <br>
-        There is a previous step: {{step.hasPrevious}}
-      </w.Step>
-
-      <w.Step as |step|>
-        <h3>Second</h3>
-        There is a first step: {{step.hasNext}} <br>
-        There is a previous step: {{step.hasPrevious}}
-      </w.Step>
-
-      <ControlGroup>
-        <button
-          type="button"
-          disabled={{not w.hasPreviousStep}}
-          {{on "click" w.transition-to-previous}}
-        >
-          Previous
-        </button>
-        <button
-          type="button"
-          disabled={{not w.hasNextStep}}
-          {{on "click" w.transition-to-next}}
-        >
-          Next
-        </button>
-      </ControlGroup>
-    </StepManager>
-  </demo.example>
-
-  <demo.snippet @name="demo-cyclical-steps-step-yielded.hbs" />
-</DocsDemo>
