--- tests/dummy/app/docs/features/validating-steps/basic-usage/component.js
@@ -1,16 +1,0 @@
-// BEGIN-SNIPPET validating-steps-basic-usage.js
-import Component from '@glimmer/component';
-import { tracked } from '@glimmer/tracking';
-import { action } from '@ember/object';
-
-export default class extends Component {
-  @tracked password = '';
-
-  @action
-  checkPassword(resolve) {
-    if (this.password === 'password') {
-      resolve();
-    }
-  }
-}
-// END-SNIPPET
