--- tests/dummy/app/docs/features/validating-steps/basic-usage/template.hbs
@@ -1,25 +1,0 @@
-<DocsDemo as |demo|>
-  <demo.example @name="validating-steps-basic-usage.hbs">
-    <StepManager as |w|>
-      <w.Step>
-        <input placeholder="Enter password" value={{this.password}}>
-
-        <ControlGroup>
-          <button
-            type="button"
-            {{on "click" (validate-transition w.transition-to-next with=this.checkPassword)}}
-          >
-            Check Password
-          </button>
-        </ControlGroup>
-      </w.Step>
-
-      <w.Step>
-        Great job! You got the password right
-      </w.Step>
-    </StepManager>
-  </demo.example>
-
-  <demo.snippet @name="validating-steps-basic-usage.hbs" />
-  <demo.snippet @name="validating-steps-basic-usage.js" @label="component.js" />
-</DocsDemo>
