--- tests/dummy/app/docs/template.hbs
@@ -1,42 +1,0 @@
-<DocsHeader />
-
-<DocsViewer as |viewer|>
-  <viewer.nav as |nav|>
-    {{nav.item "Overview" "docs.index"}}
-    {{nav.item "Cookbook" "docs.cookbook.index"}}
-
-    {{nav.section "Features"}}
-    <nav.subnav as |features|>
-      {{features.item "Validating Steps" "docs.features.validating-steps"}}
-      {{features.item "Linear & Cyclical Flow" "docs.features.state-manager"}}
-      {{features.item "Inactive State" "docs.features.inactive"}}
-    </nav.subnav>
-
-    {{nav.section "Cookbook"}}
-    <nav.subnav as |cookbook|>
-      {{cookbook.item
-        "Dynamic Definition of Steps"
-        "docs.cookbook.dynamic-definition"
-      }}
-      {{cookbook.item "Tabbed Interface" "docs.cookbook.tabs"}}
-      <nav.subnav as |wizard|>
-        {{wizard.item "Bound to Query Param" "docs.cookbook.tabs.query-param"}}
-      </nav.subnav>
-      {{cookbook.item "Wizard" "docs.cookbook.wizard"}}
-      <nav.subnav as |wizard|>
-        {{wizard.item
-          "With Progress Indicator"
-          "docs.cookbook.wizard.progress-indicator"
-        }}
-      </nav.subnav>
-    </nav.subnav>
-  </viewer.nav>
-
-  <viewer.main>
-    <div class="docs-container docs__center">
-      <div class="docs-section">
-        {{outlet}}
-      </div>
-    </div>
-  </viewer.main>
-</DocsViewer>
