+++ tests/dummy/app/routes/application.js
@@ -1,0 +1,3 @@
+import Route from '@ember/routing/route';
+
+export default class ApplicationRoute extends Route {}
