+++ tests/dummy/app/templates/application.hbs
@@ -1,0 +1,3 @@
+{{outlet}}
+
+<DocsKeyboardShortcuts />
