+++ tests/dummy/app/templates/docs/cookbook/tabs.md
@@ -1,0 +1,5 @@
+# Tabbed Interface
+
+A tabbed interface is a really common UI pattern that is often implemented using a number of conditionals in the template. You can simplify this greatly using a `step-manager` that tracks which tab is active and allows you to "transition" to a new tab.
+
+<Docs::Cookbook::Tabs::DemoBasicTabs />
