+++ tests/dummy/app/templates/docs/features/inactive.md
@@ -1,0 +1,5 @@
+# Rendering Inactive Steps
+
+In most cases you'll only want to show one step at a time. However, if you would rather display a "disabled" state for steps that are not active, you can do that too!
+
+<Docs::Features::Inactive::Example1 />
